Remove EntityManager
