+++ RedirectService.php
@@ -52,10 +52,10 @@ class RedirectService {
     }
     
     /**
-     * @return \Doctrine\ORM\EntityManager
+     * @return \Doctrine\Common\Persistence\ObjectManager
      */
-    protected function getEntityManager() {
-        return $this->doctrine->getEntityManager();
+    protected function getManager() {
+        return $this->doctrine->getManager();
     }
     
     /**
@@ -79,14 +79,9 @@ class RedirectService {
         if ($existingRoute !== false)
             throw new Exception\RouteAlreadyExistsException('Can not create redirect route "'. $slug . '" to object "'. $objectType . '" with id "'.$objectId.'", because such route already exists in "'.$locale.'" locale.');   
         
-        $em = $this->getEntityManager();
-
-        $q = $em->createQuery('SELECT r FROM ObjectRouterBundle:ObjectRoute r WHERE r.object_type = ?1 AND r.object_id = ?2 AND r.lng = ?3');
-        $q->setParameter(1, $objectType);
-        $q->setParameter(2, $objectId);
-        $q->setParameter(3, $locale);
+        $om = $this->getManager();
         
-        $objectRouteResults = $q->execute();
+        $objectRouteResults = $om->getRepository("ObjectRouterBundle:ObjectRoute")->getObjectRoutes($objectType, $objectId, $locale);
         if (count($objectRouteResults) == 0) {
             throw new Exception\ObjectHasNoRouteException('Can not create redirect route "'. $slug . '" to object "'. $objectType . '" with id "'.$objectId.'", because object has no defined route in "'.$locale.'" locale.');
         }
@@ -97,79 +92,15 @@ class RedirectService {
         $redirect->setObjectRoute($route);
         $redirect->setType($type);
         
-        $em->persist($redirect);
+        $om->persist($redirect);
         
-        $em->flush();
+        $om->flush();
         
         $this->objectRouter->setSlug('redirect', $redirect->getId(), $locale, $slug, true);
         
         return $redirect->getId();
     }
     
-    /**
-     * Get redirect links to an object.
-     * 
-     * @param string $objectType Object type
-     * @param integer $objectId Object Id
-     * @param integer $type Redirect type (301, 302, ...). If false, returns all redirects.
-     * @param string $locale Object locale. If false, returns links to object in all locales.
-     * @return array Returns array of 
-     *      "slug" - redirect route slug, 
-     *      "linkId" - redirect item id,
-     *      "locale" - redirect route locale
-     *      "visible" - true if redirect is visible
-     *      "redirectType" - redirect type (301, 302, ...)
-     *      "objectLng" - target object language for this specific route
-     */
-    public function getRedirectsToObject($objectType, $objectId, $type = false, $locale = false) {
-                
-        $em = $this->getEntityManager();
-        
-        $qb = $em->createQueryBuilder();
-        $qb->from('ObjectRouterBundle:ObjectRouteRedirect', 'rr')
-                ->innerJoin('rr.objectRoute', 'r')
-                ->andWhere('r.object_type = ?1')
-                ->andWhere('r.object_id = ?2')
-                ->select('rr.id as linkId, r.lng as objectLng, rr.type as redirectType')
-                ->setParameter(1, $objectType)
-                ->setParameter(2, $objectId);
-
-        if ($locale !== false) {
-            $qb->andWhere('r.lng = :lng')->setParameter('lng', $locale);
-        }
-        
-        if ($type !== false) {
-            $qb->andWhere('rr.type = :type')->setParameter('type', $type);
-        }
-        
-        $link_ids = array();
-        $link_data = array();
-        foreach ($qb->getQuery()->getArrayResult() as $row) {
-            $link_ids[] = $row['linkId'];
-            $link_data[$row['linkId']] = $row;
-        }
-        
-        if (empty($link_ids))
-            return array();
-        
-        $qb = $em->createQueryBuilder();
-        $qb->from('ObjectRouterBundle:ObjectRoute', 'r')
-                ->andWhere('r.object_type = :ot')
-                ->andWhere('r.object_id IN (:ids)')
-                ->select('r.slug, r.object_id as linkId, r.lng as locale, r.visible')
-                ->setParameter('ot', 'redirect')
-                ->setParameter('ids', $link_ids);
-               
-        $results = $qb->getQuery()->execute();
-        
-        foreach ($results as &$row) {
-            $row['redirectType'] = $link_data[$row['linkId']]['redirectType'];
-            $row['objectLng'] = $link_data[$row['linkId']]['objectLng'];
-        }
-        
-        return $results;
-        
-    }
     
     public function updateRedirect($linkId, $link, $type) {
         
@@ -186,12 +117,9 @@ class RedirectService {
      * @return mixed Boolean FALSE if route slug was not found or array(lng, slug, type) 
      */
     public function getLinkLocaleSlugAndType($linkId) {
-        $em = $this->getEntityManager();
+        $om = $this->getManager();
         
-        $q = $em->createQuery('SELECT r.lng, r.slug, rr.type FROM ObjectRouterBundle:ObjectRouteRedirect rr INNER JOIN rr.objectRoute r WHERE rr.id = ?1');
-        $q->setParameter(1, $linkId);
-        
-        $results = $q->execute();
+        $results = $om->getRepository("ObjectRouterBundle:ObjectRouteRedirect")->getLinkLocaleSlugAndType($linkId);
         
         if (count($results) == 0)
             return false;
