+++ RoutingEditHelper.php
@@ -11,44 +11,42 @@ namespace Nercury\ObjectRouterBundle;
 class RoutingEditHelper {
 
     /**
-     * @var EntityManager
+     * @var \Doctrine\Common\Persistence\ObjectManager
      */
-    private $em;
+    private $om;
     private $objectType;
     private $objectId;
 
-    public function __construct($em, $objectType, $objectId) {
-        $this->em = $em;
+    public function __construct($om, $objectType, $objectId) {
+        $this->om = $om;
         $this->objectType = $objectType;
         $this->objectId = $objectId;
     }
 
     public function getRoutingItems() {
-        return $this->em
-            ->createQuery('SELECT r FROM ObjectRouterBundle:ObjectRoute r WHERE r.object_type = :object_type AND r.object_id = :object_id')
-            ->setParameter('object_type', $this->objectType)
-            ->setParameter('object_id', $this->objectId)
-            ->getResult();
+        $items = $this->om->getRepository('ObjectRouterBundle:ObjectRoute')->getRoutingItems($this->objectType, $this->objectId);
+        
+        return $items;
     }
 
     public function updateRoutingItems(GeneratorService $routerGenerator, &$items) {
         foreach ($items as $k => &$rItem) {
             if ($rItem->getSlug() === null || trim($rItem->getSlug()) === '') {
                 unset($items[$k]);
-                if ($this->em->contains($rItem)) {
-                    $this->em->remove($rItem);
+                if ($this->om->contains($rItem)) {
+                    $this->om->remove($rItem);
                 }
                 continue;
             }
 
-            if (!$this->em->contains($rItem)) {
+            if (!$this->om->contains($rItem)) {
                 $rItem->setObjectType($this->objectType);
                 $rItem->setObjectId($this->objectId);
                 $rItem->setVisible(false);
                 $rItem->setSlug($routerGenerator->generateUniqueSlug('product', $this->objectId, $rItem->getLng(), $rItem->getSlug()));
-                $this->em->persist($rItem);
+                $this->om->persist($rItem);
             } else {
-                $uow = $this->em->getUnitOfWork();
+                $uow = $this->om->getUnitOfWork();
                 $uow->computeChangeSets();
                 $changeset = $uow->getEntityChangeSet($rItem);
                 if (isset($changeset['slug'])) {
